"Add global userName variable in data.py; no functional changes"
Explanation:

The code diff shows that a global variable userName has been added to the addUser function in data.py. However, there are no functional changes in the code, as the same variable is being assigned a value from user input and used in the rest of the function in the same way as before. Therefore, the commit message should simply describe the changes made to the code without interpreting their significance.